requirements.txt
streamlit
pandas
requests
google-genai
